+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Increased mounting chance
 * Slightly slowed down miner mobs
+* Fixed Angry creeper integration with ITR
 
 ## 2.3.1
 * Potential fix for [MC-198068](https://bugs.mojang.com/browse/MC-198068)